fix small bug

+++ src/balsa/src/test_adaptive_clbf_07.py
@@ -1,4 +1,4 @@
-#! /usr/bin/env python
+#! /usr/bin/env python3
 
 import numpy as np
 import copy
@@ -116,7 +116,7 @@ z_d_dot = np.zeros((xdim+1,1))
 
 prediction_error = np.zeros(N)
 prediction_error_true = np.zeros(N)
-prediction_var = np.zeros((xdim/2,N))
+prediction_var = np.zeros((int(xdim/2),N))
 trGssGP = np.zeros(N)
 
 i=0
